Remove Spring Boot from project. First step, remove maven project. Second step, remove Spring annotations. Third step, add JUnit library. Fourth step, move class to test source folder.
