--- src/main/java/com/idorigem/softplan/SoftplanApplication.java
@@ -1,13 +1,0 @@
-package com.idorigem.softplan;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SoftplanApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SoftplanApplication.class, args);
-	}
-
-}
