--- src/test/java/com/idorigem/softplan/SoftplanApplicationTests.java
@@ -1,13 +1,0 @@
-package com.idorigem.softplan;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SoftplanApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
